+++ test/commands/dev/index.test.ts
@@ -8,8 +8,10 @@ import AppService from '../../../src/app-config/service';
 import AccountUtils from '../../../src/architect/account/account.utils';
 import SecretUtils from '../../../src/architect/secret/secret.utils';
 import Dev, { UpProcessManager } from '../../../src/commands/dev';
+import { DockerUtils } from '../../../src/common/docker';
 import { DockerComposeUtils } from '../../../src/common/docker-compose';
 import DockerComposeTemplate from '../../../src/common/docker-compose/template';
+import { DockerHelper } from '../../../src/common/docker/helper';
 import PluginManager from '../../../src/common/plugins/plugin-manager';
 import DeployUtils from '../../../src/common/utils/deploy.utils';
 import * as ComponentBuilder from '../../../src/dependency-manager/spec/utils/component-builder';
@@ -367,7 +369,7 @@ describe('local dev environment', function () {
         labels: ['architect.ref=database-seeding.services.my-demo-db'],
       },
       'gateway': {
-        'image': 'traefik:v2.6.2',
+        'image': 'traefik:v2.9.8',
         'command': [
           '--api.insecure=true',
           '--pilot.dashboard=false',
@@ -376,6 +378,7 @@ describe('local dev environment', function () {
           '--accesslog.filters.statusCodes=400-599',
           '--entryPoints.web.address=:80',
           '--providers.docker=true',
+          '--providers.docker.allowEmptyServices=true',
           '--providers.docker.exposedByDefault=false',
           '--providers.docker.constraints=Label(`traefik.port`,`80`)',
         ],
@@ -424,7 +427,7 @@ describe('local dev environment', function () {
         },
       },
       'gateway': {
-        'image': 'traefik:v2.6.2',
+        'image': 'traefik:v2.9.8',
         'command': [
           '--api.insecure=true',
           '--pilot.dashboard=false',
@@ -433,6 +436,7 @@ describe('local dev environment', function () {
           '--accesslog.filters.statusCodes=400-599',
           '--entryPoints.web.address=:80',
           '--providers.docker=true',
+          '--providers.docker.allowEmptyServices=true',
           '--providers.docker.exposedByDefault=false',
           '--providers.docker.constraints=Label(`traefik.port`,`80`)',
         ],
@@ -481,7 +485,7 @@ describe('local dev environment', function () {
         },
       },
       'gateway': {
-        'image': 'traefik:v2.6.2',
+        'image': 'traefik:v2.9.8',
         'command': [
           '--api.insecure=true',
           '--pilot.dashboard=false',
@@ -490,6 +494,7 @@ describe('local dev environment', function () {
           '--accesslog.filters.statusCodes=400-599',
           '--entryPoints.web.address=:443',
           '--providers.docker=true',
+          '--providers.docker.allowEmptyServices=true',
           '--providers.docker.exposedByDefault=false',
           '--providers.docker.constraints=Label(`traefik.port`,`443`)',
           '--serversTransport.insecureSkipVerify=true',
@@ -537,7 +542,7 @@ describe('local dev environment', function () {
         "external_links": [
           "gateway:hello.arc.localhost"
         ],
-        "image": "hello-world--api:latest",
+        "image": "hello-world--api",
         "healthcheck": {
           "test": [
             "CMD", "curl", "--fail", "localhost:3000"
@@ -549,7 +554,7 @@ describe('local dev environment', function () {
         },
       },
       "gateway": {
-        "image": "traefik:v2.6.2",
+        "image": "traefik:v2.9.8",
         "command": [
           "--api.insecure=true",
           "--pilot.dashboard=false",
@@ -558,6 +563,7 @@ describe('local dev environment', function () {
           "--accesslog.filters.statusCodes=400-599",
           "--entryPoints.web.address=:80",
           "--providers.docker=true",
+          "--providers.docker.allowEmptyServices=true",
           "--providers.docker.exposedByDefault=false",
           "--providers.docker.constraints=Label(`traefik.port`,`80`)",
         ],
@@ -595,7 +601,7 @@ describe('local dev environment', function () {
           "gateway:buildpack-api.arc.localhost",
           "gateway:dockerfile-api.arc.localhost"
         ],
-        "image": "hello-world--buildpack-api:latest",
+        "image": "hello-world--buildpack-api",
         "healthcheck": {
           "test": [
             "CMD", "curl", "--fail", "localhost:3000"
@@ -608,7 +614,11 @@ describe('local dev environment', function () {
       },
       "hello-world--dockerfile-api": {
         "build": {
-          "context": path.resolve("./test/integration/hello-world")
+          "context": path.resolve("./test/integration/hello-world"),
+          "tags": [
+            "hello-world--dockerfile-api",
+            "hello-world--dockerfile-api2"
+          ]
         },
         "ports": [
           "50001:4000",
@@ -639,8 +649,30 @@ describe('local dev environment', function () {
           "start_period": "0s"
         },
       },
+      "hello-world--dockerfile-api2": {
+        "environment": {},
+        "external_links": [
+          "gateway:buildpack-api.arc.localhost",
+          "gateway:dockerfile-api.arc.localhost"
+        ],
+        "image": "hello-world--dockerfile-api2",
+        "labels": [
+          "architect.ref=hello-world.services.dockerfile-api2"
+        ]
+      },
+      "hello-world--redis": {
+        "environment": {},
+        "external_links": [
+          "gateway:buildpack-api.arc.localhost",
+          "gateway:dockerfile-api.arc.localhost"
+        ],
+        "image": "redis",
+        "labels": [
+          "architect.ref=hello-world.services.redis"
+        ]
+      },
       "gateway": {
-        "image": "traefik:v2.6.2",
+        "image": "traefik:v2.9.8",
         "command": [
           "--api.insecure=true",
           "--pilot.dashboard=false",
@@ -649,6 +681,7 @@ describe('local dev environment', function () {
           "--accesslog.filters.statusCodes=400-599",
           "--entryPoints.web.address=:80",
           "--providers.docker=true",
+          "--providers.docker.allowEmptyServices=true",
           "--providers.docker.exposedByDefault=false",
           "--providers.docker.constraints=Label(`traefik.port`,`80`)",
         ],
@@ -764,7 +797,12 @@ describe('local dev environment', function () {
     .timeout(20000)
     .stub(ComponentBuilder, 'buildSpecFromPath', () => {
       const spec = buildSpecFromYml(yaml.dump(local_database_seeding_component_config));
-      spec.metadata.file = { path: getMockComponentFilePath('database-seeding'), contents: '' };
+      const component_path = getMockComponentFilePath('database-seeding');
+      spec.metadata.file = {
+        path: component_path,
+        folder: fs.lstatSync(component_path).isFile() ? path.dirname(component_path) : component_path,
+        contents: ''
+      };
       return spec;
     })
     .stub(Dev.prototype, 'failIfEnvironmentExists', sinon.stub().returns(undefined))
@@ -1356,6 +1394,8 @@ describe('local dev environment', function () {
     .stub(PluginManager, 'getPlugin', sinon.stub().returns({
       build: () => { },
     }))
+    .stub(DockerHelper, 'composeVersion', sinon.stub().returns(true))
+    .stub(DockerHelper, 'buildXVersion', sinon.stub().returns(true))
     .stdout({ print })
     .stderr({ print })
     .command(['dev', './test/mocks/buildpack/buildpack-dockerfile-architect.yml', '--ssl=false'])
@@ -1370,6 +1410,7 @@ describe('local dev environment', function () {
     .stub(ComponentBuilder, 'loadFile', () => {
       return fs.readFileSync('test/mocks/register/nonexistence-dockerfile-architect.yml').toString();
     })
+    .stub(DockerUtils, 'doesDockerfileExist', sinon.stub().callsFake(DockerUtils.doesDockerfileExist)) // override global stub
     .stub(Dev.prototype, 'failIfEnvironmentExists', sinon.stub().returns(undefined))
     .stub(Dev.prototype, 'runCompose', sinon.stub().returns(undefined))
     .stub(Dev.prototype, 'downloadSSLCerts', sinon.stub().returns(undefined))